08.24 12:25 finish notice etc
