+++ app/src/main/res/drawable-hdpi/noticeitem_border.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_height="match_parent"
+    android:layout_width="match_parent">
+
+
+
+    <item>
+
+        <shape android:shape="rectangle">
+
+            <solid android:color="#d7d7d7" />
+
+        </shape>
+
+    </item>
+
+    <item android:top="0dp" android:bottom="2dp">
+
+        <shape android:shape="rectangle">
+
+            <solid android:color="#ffffff" />
+
+        </shape>
+
+    </item>
+
+
+
+</layer-list>
