+++ app/src/main/res/layout/etc_qna_fragment.xml
@@ -7,6 +7,4 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-
-
 </LinearLayout>
