+++ app/src/main/res/layout/noticeitem.xml
@@ -7,7 +7,7 @@
     android:padding="8dp"
     android:gravity="center_vertical"
     android:layout_marginBottom="8dp"
-    android:background="@android:color/white">
+    android:background="@drawable/noticeitem_border">
 
     <TextView
         android:id="@+id/textView1"
